% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bangladesh}
\alias{Bangladesh}
\title{Bangladesh}
\format{
A data frame with 271 rows and 3 variables:
}
\usage{
Bangladesh
}
\description{
The Bangladesh data is provided solely for illustrative purposes and to
enable statistical analysis.  Full data are available from the British
Geological Survey web site:
\url{https://www2.bgs.ac.uk/groundwater/health/arsenic/Bangladesh/home.html}.
}
\details{
A data set containing 271 ??. The variables are as follows:
}
\keyword{datasets}
