% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Basketball09}
\alias{Basketball09}
\title{Basketball09}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Basketball09
}
\description{
A dataset containing X. The variables are as follows:
}
\keyword{datasets}

