% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Batters2005}
\alias{Batters2005}
\title{Batters2005}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Batters2005
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}

