% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BookPrices}
\alias{BookPrices}
\title{Price of hardcover textbooks from mathematics and the natural sciences and
the social sciences.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
BookPrices
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
