% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bushmeat2}
\alias{Bushmeat2}
\title{Bushmeat in Ghana.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Bushmeat2
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}

