% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cereals}
\alias{Cereals}
\title{Information on various cereals.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Cereals
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
