% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Challenger}
\alias{Challenger}
\title{Data on 23 Challenger flights.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Challenger
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}

