% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corrExerciseA}
\alias{corrExerciseA}
\title{Correlation Exercise A p.294}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
corrExerciseA
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}

