% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corrExerciseB}
\alias{corrExerciseB}
\title{Correlation Exercise B p.294}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
corrExerciseB
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}

