% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Diving2017}
\alias{Diving2017}
\title{Diving times in 2017}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Diving2017
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
