% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fatalities}
\alias{Fatalities}
\title{Fatalities}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Fatalities
}
\description{
A dataset containing X. The variables are as follows:
}
\keyword{datasets}

