% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FishMercury}
\alias{FishMercury}
\title{Mercury levels (parts per million) for 30 fish caught in lakes in Minnesota.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
FishMercury
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
