% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Girls2004}
\alias{Girls2004}
\title{Girls2004}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Girls2004
}
\description{
A dataset containing X. The variables are as follows:
}
\keyword{datasets}

