% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSS2002}
\alias{GSS2002}
\title{Results from 2002 General Society Survey.}
\format{
A data frame with X rows and Y variables:
}
\usage{
GSS2002
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
