% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSS2006}
\alias{GSS2006}
\title{Results from 2006 General Society Survey.}
\format{
A data frame with X rows and Y variables:
}
\usage{
GSS2006
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
