% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IceCream}
\alias{IceCream}
\title{Calorie information for a sample of brands of chocolate and vanilla ice cream.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
IceCream
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}

