% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Maunaloa}
\alias{Maunaloa}
\title{Data on average CO2 levels (ppm) for the month of May from 1990 to 2010.}
\format{
A data frame with X rows and Y variables:
}
\usage{
Maunaloa
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
