% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA1617}
\alias{NBA1617}
\title{NBA 2016-2017 data.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
NBA1617
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
