% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Olympics2012}
\alias{Olympics2012}
\title{2012 Olympics Data.}
\format{
A data frame with X rows and Y variables:
}
\usage{
Olympics2012
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
