% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Phillies2009}
\alias{Phillies2009}
\title{Data from the 2009 season for the baseball team of the Philadelphia Phillies.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Phillies2009
}
\description{
A dataset containing X. The variables are as follows:
}
\keyword{datasets}

