% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Quetzal}
\alias{Quetzal}
\title{Quetzal.}
\format{
A data frame with X rows and Y variables:
}
\usage{
Quetzal
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
