% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RangersTwins2016}
\alias{RangersTwins2016}
\title{Rangers/Twins 2016 Baseball Data.}
\format{
A data frame with X rows and Y variables:
}
\usage{
RangersTwins2016
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
