% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampledata.R
\docType{package}
\name{resampledata-package}
\alias{resampledata-package}
\title{resampledata: Data Sets for Mathematical Statistics with Resampling in R}
\description{
Package of data sets from "Mathematical Statistics with Resampling in R" (1st Ed. 2011, 2nd Ed. 2018) by Laura Chihara and Tim Hesterberg.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rudeboybert/resampledata}
  \item Report bugs at \url{https://github.com/rudeboybert/resampledata/issues}
}

}
\author{
\strong{Maintainer}: Albert Y. Kim \email{albert.ys.kim@gmail.com} (\href{https://orcid.org/0000-0001-7824-306X}{ORCID})

Authors:
\itemize{
  \item Laura Chihara \email{lchihara@carleton.edu}
  \item Tim Hesterberg \email{TimHesterberg@gmail.com}
}

}
\keyword{internal}
