% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampledata.R
\docType{package}
\name{resampledata}
\alias{resampledata}
\alias{resampledata-package}
\title{resampledata: Data Sets for Mathematical Statistics with Resampling in R}
\description{
Package of data sets from "Mathematical Statistics with Resampling in R"
(2011) by Laura Chihara and Tim Hesterberg.
}
\section{Data sets}{

Type \code{data(package="resampledata")} to view all datasets. Note the
following datasets were not included due to sharing restrictions and can only
 be accessed on the textbook's webpage: \code{Bangladesh}, \code{GSS2002},
 \code{GSS2006}, and \code{Volleyball2009}.
}
\examples{
library(resampledata)
data(FlightDelays)
}

