% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sat2008}
\alias{Sat2008}
\title{Sat2008}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Sat2008
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}

