% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Skateboard}
\alias{Skateboard}
\title{Skateboarding data}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Skateboard
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
