% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Skating2010}
\alias{Skating2010}
\title{Skating2010}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Skating2010
}
\description{
A dataset containing X. The variables are as follows:
}
\keyword{datasets}

