% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Spruce}
\alias{Spruce}
\title{Study of factors affecting the growth of the black spruce.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Spruce
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
