% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Titanic}
\alias{Titanic}
\title{Data on male passengers on the Titanic.}
\format{
A data frame with X rows and Y variables:
}
\usage{
Titanic
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
