% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Turbine}
\alias{Turbine}
\title{Wind Speeds (m/s) from Carleton College Turbine.}
\format{
A data frame with X rows and Y variables:
}
\usage{
Turbine
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
