% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TV}
\alias{TV}
\title{TV}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
TV
}
\description{
A dataset containing X. The variables are as follows:
}
\keyword{datasets}

