% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Verizon}
\alias{Verizon}
\title{Random sample of repair times for 1664 ILEC and 23 CLEC customers.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Verizon
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
