% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{volleyball2005}
\alias{volleyball2005}
\title{volleyball2005}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
volleyball2005
}
\description{
A dataset containing X. The variables are as follows:
}
\keyword{datasets}

