% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Volleyball2009}
\alias{Volleyball2009}
\title{Volleyball2009}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Volleyball2009
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
