% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wafers}
\alias{wafers}
\title{Wafers}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
wafers
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
