% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Walleye}
\alias{Walleye}
\title{Length and weight measurements for a sample of 60 walleye caught in Minnesota lakes.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Walleye
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}

