% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Alelager}
\alias{Alelager}
\title{Calories and alcohol content for ales and lagers.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Alelager
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}

