% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fatalities}
\alias{Fatalities}
\title{Random sample of 100 driver fatalities in 2009 in Pennsylvania.}
\format{
A data frame with X rows and Y variables:
}
\usage{
Fatalities
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
