% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSS2006}
\alias{GSS2006}
\title{GSS2006}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
GSS2006
}
\description{
A dataset containing X. The variables are as follows:
}
\keyword{datasets}

