% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Illiteracy}
\alias{Illiteracy}
\title{Data on a sample of countries where female illiteracy is more that 5 percent.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Illiteracy
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
