% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lottery}
\alias{Lottery}
\title{Winning numbers for the daily games from May 5, 2010 through August 15, 2010.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Lottery
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}

