% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MnGroundwater}
\alias{MnGroundwater}
\title{MnGroundwater}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
MnGroundwater
}
\description{
A dataset containing X. The variables are as follows:
}
\keyword{datasets}

