% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCBirths2004}
\alias{NCBirths2004}
\title{Random sample of 1009 babies born in North Carolina during 2004.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
NCBirths2004
}
\description{
A dataset containing X. The variables are as follows:
}
\keyword{datasets}

