% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampledata.R
\docType{package}
\name{resampledata}
\alias{resampledata}
\alias{resampledata-package}
\title{resampledata: Data Sets for Mathematical Statistics with Resampling in R}
\description{
Package of data sets from "Mathematical Statistics with Resampling in R"
(2011) by Laura Chihara and Tim Hesterberg.
}
\section{Data sets}{

Type \code{data(package="resampledata")} to view all datasets.
}
\examples{
library(resampledata)
data(FlightDelays)
}

