% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampledata.R
\docType{package}
\name{resampledata}
\alias{resampledata}
\alias{resampledata-package}
\title{resampledata: Datasets for Mathematical Statistics with Resampling in R}
\description{
Package to load datasets from "Datasets for Mathematical Statistics
with Resampling in R" (2011) by Laura Chihara and Tim Hesterberg.
}
\section{Foo functions}{

The foo functions ...
}

