% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Skating2010}
\alias{Skating2010}
\title{Scores from the short program and free skate for men's figure skating in the 2010 Olympics.}
\format{
A data frame with X rows and Y variables:
}
\usage{
Skating2010
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
