% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TXBirths2004}
\alias{TXBirths2004}
\title{Random sample of 1587 babies born in Texas in 2004.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
TXBirths2004
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}

