% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Verizon}
\alias{Verizon}
\title{Verizon}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Verizon
}
\description{
A dataset containing X. The variables are as follows:
}
\keyword{datasets}

