% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Walleye}
\alias{Walleye}
\title{Walleye}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Walleye
}
\description{
A dataset containing X. The variables are as follows:
}
\keyword{datasets}

