% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Watertable}
\alias{Watertable}
\title{Relationship between seedling growth and water table depth for a sample of seedlings.}
\format{A data frame with X rows and Y variables:
\itemize{
  \item
}}
\usage{
Watertable
}
\description{
A data set containing X. The variables are as follows:
}
\keyword{datasets}
